refactor(contracts/tasks): Extract tx sending logic in a separate function
